+++ packages/widgetbook/lib/src/core/component.dart
@@ -1,13 +1,14 @@
 import 'package:flutter/widgets.dart';
 import 'package:path/path.dart' as p;
 
-import 'meta.dart';
 import 'story.dart';
 import 'story_args.dart';
 
 class Component<TWidget extends Widget, TArgs extends StoryArgs<TWidget>> {
   Component({
-    required this.meta,
+    required this.name,
+    this.path = '',
+    this.docs,
     required this.stories,
   }) {
     stories.forEach((story) {
@@ -15,14 +16,12 @@ class Component<TWidget extends Widget, TArgs extends StoryArgs<TWidget>> {
     });
   }
 
-  final Meta<TWidget> meta;
+  final String name;
+  final String path;
+  final String? docs;
   final List<Story<TWidget, TArgs>> stories;
 
-  String get name => meta.name;
-  String? get path => meta.path;
-  String? get docs => meta.docs;
-
-  String get fullPath => p.join(path ?? '', name);
+  String get fullPath => p.join(path, name);
   String get docsPath => p.join(fullPath, 'Docs');
 
   String pathOf(Story story) {