+++ packages/widgetbook/lib/src/state/widgetbook_state.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:meta/meta.dart';
 
-import '../../next.dart';
 import '../addons/addons.dart';
 import '../fields/fields.dart';
 import '../integrations/widgetbook_integration.dart';
@@ -295,16 +294,6 @@ class WidgetbookState extends ChangeNotifier {
 
     notifyListeners();
   }
-
-  /* Widgetbook Next: SAM (Story-Arg-Mode) Structure */
-
-  /// Returns `true` if SAM (Story-Arg-Mode) structure is used.
-  @experimental
-  bool get isNext => useCase is Story;
-
-  /// Returns the current active [Story].
-  @experimental
-  Story? get story => isNext ? useCase as Story : null;
 
   @override
   void dispose() {