+++ packages/widgetbook/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 3.18.0
 
 - **BREAKING**: Remove `next` library. ([#1634](https://github.com/widgetbook/widgetbook/pull/1634))